requirements.txt
async-timeout==4.0.3
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.2.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.4
numpy==1.25.1
pandas==2.0.3
polars==0.18.6
protobuf==4.25.1
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
spotipy==2.23.0
tzdata==2023.3
urllib3==2.0.3
